model vattributes: fix remote unsubscribing causing all remote listeners to stop listening
Each remote listener was subscribing with the _same_ ID.
So the VA thought there was just a single remote listener always.
That meant that as soon as a remote listener would unsubscribe, all the other remote
listeners would stop receiving updates... until one remote subscriber happens again.

In practice, this surprisingly didn't cause too much issues, but did
have very hidden issues such as the polePosition not always updating the metadata
of the CCD (on the SPARC).

=> Do exactly the same as on the DataFlow: have a special ID per VA proxy.