acq align delphi_man_calib: use explicit ANSI colours
Create special print and input functions that support the ANSI colour codes.
This avoid have the unreadable colour codes in each message string.

Also display the error in case of unexcepted exception during the run.
Also wait for the user to press enter before fully ending. That's handy
because the calibration is typically called in a separate terminal which
automatically closes when the calibration is over, so that user wouldn't
be able to see the error message if it ever happens.