gui cont microscope: fix the optical focus not referenced on the first load
Until now it referenced before loading the calibration file.
That was all fine until the referencing of the optical focus depended on the
calibration file.

This would mean that the first time the sample is load, the optical focus
is considered not known, and so it's not used.

=> Load the calibration first.

Also add a warning if the optical focus is (really) not known, as nowadays
all the Delphis should be calibrated with the optical focus.