gui util: Clean-up decorators
There are decorators and wrappers. They do the same thing, so let's make sure they
have the same code (modulo the fact that decorators are for wx.Object methods).

Also handle better when a widget is destroyed and has wxlimit_invocation().
With wxlimit_invocation, the wrapper cannot check if the object is dead before calling the function.
So explicitly check for the object being dead in the decorated functions.

Also unsubscribe from the most obvious events when a Viewport is destroyed.
That ensures that it doesn't keep processing events long after it's not in use.