driver static lens: check pole position is a number
Until now, we checked that the pole position value was > 0.
However, in Python any string is larger than a number.
So, if the user enters the pole position mistakenly as string, it will be considered fine.
=> Also check it's smaller than 1 million.
This way, we check that the number is not wrong, and that it's a number.