driver semcomedi: move the ebeam to the first pixel ASAP
As it can take a long time to actually start writting the scan position values,
it's nice to get an extra early position for the very first pixel.
That pixel is a little special as the e-beam has to move either from the
park position (on a first scan) other end of the scan (on a second+ scan).