gui cont tabs: remove unsused code
One day, we thought it could come in handy later on.
A few years later, we still haven't fixed properly showing multiple spectrum
streams at the same time, but I'm get convinced that if we ever fix it,
it will be with code in comp/viewports. So this code is really not helpful.
=> Remove