gui cont streams: fix repetition not being displayed
Another fallout from commit d690dea0a957 (Clean-up decorators).
Now the update to the value of the combobox would happen sooner.
Even sooner than the change of the choices of the combobox,
which so far had always happened before due to total luck.

So this change broke our luck. Now we have to be careful, and explicitly
put the value after clearing the combobox, because no one is doing it for us.

Also make sure that extra choice update happens in the main GUI thread,
as it touches the widgets. In practice, so far, it always happened in the main
GUI thread anyway.