driver actuator ConvertStage: don't fail if moveRel() or moveAbs() is called with update
If the underlying actuator supports canUpdate, then this flag is passed as is
on the axis definition. So at least we need to either support the "update" argument,
or drop the canUpdate. We pick the former solution.

To be actually correct, if some rotation is applied,
the canUpdate should only be set if both axes canUpdate.
For now that not the case, anyway, in practice all the actuators
which support canUpdate on an axis can also do it on the other axis.