acq stream SEMCCDMDStream: increase time between spot positionning and CCD exposure
We used to almost immediately start the CCD after moving the spot.
However, on very short exposure time (~10ms) spectral acquistion, it was possible to
notice incorrect data for some spots (10%).
Adding a small delay of 10ms seems to avoid the problem.
This is not the proper way, and there are much better solutions, but at least,
it's a easy fix, which can easily be reverted once we have implemented better ways.

back-port from 12e0245cf87efe5462b604f8e83f964cad70d061