gui cont tabs: Fix SEM- and optical-only microscope showing image as a chronogram
With the addition of the chronogram view for the confocal FLIM support,
the generic to handle just SEM or just optical got "too" generic.
In such case, the stream shouldn't be assigned to the chronogram view.

This avoids errors when playing stream, as the chronogram doesn't handle a standard 2D image.