gui comp: fix white pixel around combobox button
It seems that in wxWidget, OwnerDrawnComboBox.GetBackgroundColour()
doesn't return the value from SetBackgroundColour(). Not sure since when.
Anyway, that caused the background of the combobox button to be white.

Also simplify buttons._create_bitmap() to get an wxImage as input instead of wxBitmap.
Previously, all the callers first got a wxImage, and converted it to wxBitmap just for passing
it to _create_bitmap(). => This avoids a wxBitmap creation.

Also add a test case for the combo module.