gui cont tabs: fix some pyramidal images sometimes not showing up when opening them
To refit the view content we used canvas.fit_view_to_next_image = True.
That works fine with standard streams. However, with pyramidal streams,
the view will _first_ set it's current rect+mpp and only after that the
canvas will be refitted. The problem is that if the current view rect+mpp
is outside of the data, the pyramidal stream has _no_ image.
If there isn't any image, the canvas never refit on the "next" image.

=> Explicitely ask to refit the canvas, as soon as the streams are inserted.
It used to not work well, because it used to rely on the current image displayed,
which could have not yet been passed to the canvas. However, the canvas now explicitly
asks the streams, so that now works fine.