scripts: improve analyse_shifts, and make combinatorial_acq acquire more combinations
To improve the accuracy of the shift reported by analyse_shifts, we always scale _up_ the smaller image,
instead of sometimes scaling down the big image.

Also acquire more data for the combinatorial acquisition, and end
with an image with the same settings as at the beginning, to check the drift.