driver spectrometer: do software binning if hardware doesn't support full vertical binning
If the underlying CCD doesn't support full vertical binning, but instead only allow some (typically small)
binning, when using the CCD as a spectrometer, a lot of signal would be lost, as only the center line is used.

=> In such case, acquire a multiple line image, and do software binning on the fly.
It's actually quite fast, as the processor is fast and there aren't that many pixels to sum.
The trick to do a proper binning is to substract the baseline and to not overflow the sum, but instead
clip it to the maximum value that the type supports.