gui util call_in_wx_main: fix causing sometimes anachronisms
We tried to make it clever, and only run asynchronously if it's not called
from the main thread, but in some (rare) cases, that can cause anachronic issues.
For example:
1. Call from another thread -> queued for later
2. Call from main thread -> immediately executed
=> Call 2 is executed before call 1, which could mean that an old value
is displayed on the GUI.

This was actually encountered on the tile plugin.

=> Just don't try to be clever, and always call the function asynchronously.
It's theoretically making the GUI reacting slightly slower, but in practice that's unnoticeable.
Correctness is better.

Also add a test cases, and fix the test case for wxlimit_invocation, which actually didn't test anything.