plugins tileacq: use VA setter for limiting nx/ny
Instead of using a subscriber, use a setter to limit the value.
It kind of works with a subscriber, if the code is careful to avoid
constantly modifying the value as the subscriber will not be called again
if the value stays the same.
However, it's much clearer, and safer to use a setter.

No change in behaviour.