Memory leak fix (caused by VAs never garbage collected)
Otherwise, each VA (without explicit setter) would hold a reference to self.__default_setter, which holds a reference to the VA. So this causes a cyclic reference. So the VA could only be deleted via the the garbage collector. However in Python 2, the garbage collector doesn't dare to touch it, because the VA has a __del__ method.
=> make sure we don't hold a reference by using a WeakMethod even for the default_setter.