driver tmcm: support relative encoder
It's pretty much the same as absolute encoders, but doesn't claim to be
referenced at init.
We now reset the "actual" position to the one read by the encoder so that
if there is encoder deviation check, the error doesn't accumulate.

Also move the param_file application before reading the speed/accel,
so the driver knows the right values.