dataio tiff: be more picky when to merge DataArrays together
Extend the logic, to be the same one as for the HDF5.
We pretty much _never_ want to merge DataArrays, excepted for the Fluorescence acquisition.

In practice, AR polarization acquisition with a filter wheel (ie, with MD_OUT_WL) would be
merged.

Also update the test cases to check with this metadata.