gui comp viewport/miccanvas: only listen to destroy events about themselves
The components listen to destroy events in order to call a destructor when
the GUI doesn't need them anymore.
However, by default all the events from the children are also listen to.

So when a button in a legend in a viewport got destroyed (for the polarization),
the viewport thought we didn't need it anymore (even if we still loved it).

=> Only listen to destroy events specifically about the widget itself.

Fixes a bug that caused the views from not reorganising after a AR polarization
data was opened and closed.