gui comp buttons: fix sizer not always properly fitting ImageButtons
With wxPython4, it seems that the sizer cares about the SIZE event.
ImageButton.OnSize() is an override of GenButton, which helps the sizer to
know about the size change.
=> we need to call the super(), which is a good practice typically anyway.

Also ensure the bitmap is always computed.
In some (rare) cases, the button wouldn't need a size update, which would prevent it
from displaying correctly at init.
=> compute the default image at init.

This fixes some plugins which would sometimes (on the second display), not properly
position the main buttons.