acq stream: fix drift correction on multiple streams
When the drift corrector is used for multiple streams, it uses
the data acquired during the previous stream for the next stream.
However, for now, the stream changed the shape of the drift corrector,
which broke the drift correction for the next stream.
=> Don't change the drift corrector data.

Also, if the drift corrector fail at init, it's not run, and so
it should also not be attempted to read its data at the end.
This would cause an extra error, after the end of the acquisition.
=> Don't try to read the drift corrector data if it has been disabled.