acq stream RepetitionStream: attempt to keep the request ROI even when the hardware limits the possible repetition
Until now, when changing the ROI we were adjusting the repetition, so that
the pixel size stays the same.
However, on some hardware (eg, confocal), the repetition can only be a power of two.
In this case, the pixel size would be prefered over the ROI, which would mean
the ROI was reverted. It would be impossible to do some adjustments to the ROI.

=> if the hardware modifies the expected repetition, adapt the pixel size
so that the accepted ROI is as close as possible from the requested one.

Also handle the hardware with a scale range < 1.
In such case, the minimum pixel size is much smaller.