gui canvas: do not round the coordinates to a pixel in phys_to_buffer_pos()
As the returned value represents pixels, it seemed logical to round them to integers.

However, in all the cases we actually use this function, we don't need the rounding.
Futhermore, the rounding was even causing issues for the ROA selection, as conversion
back and forth with the relative ROA would cause a change in all the values,
which would let the RepetitionStream think that the user is attempting to change
both dimensions of the ROA when he's actually only moving it around.

=> Don't round the values. If one day someone needs the pixel coordinates
rounded to integers (not Cairo!), then he'll have to do it explicitly.