acq stream: Stream.raw now contains the DataArrayShadow if pyramidal data
We used to the special ._das in case the raw data was a DataArrayShadow.
Then .raw would be a 2D tuple of the latest tiles projected.

The main issue is that if there were multiple projections (ie, views) of the same
stream, .raw would pretty much be random. It worked most of the time, though,
because it was only used by the exporter, and typically the user changes
last the view he wants to export.

=> Now say that .raw is whatever is the original data.
It's up to the user to check whether raw is actually a DataArray or a DataArrayShadow.

Now the projection provides a different .raw in case the stream has a DataArrayShadow:
a 2D tuple of the raw tiles.

It turns out to also simplify quite a bit the code.