driver andorshrk: ensure it's not possible to interleave calls to getPixelToWavelength()
In case getPixelToWavelength() is called simultaneously with info from 2 different CCDs,
it could end-up return the wrong data. That's more likely to happen than it sounds like,
because this is typically called in parallel for all the CCDs when the wavelength changes.

As the shamrock driver takes the input of GetCalibration() from previous calls (SetNumberPixels(),
SetPixelWidth()), it's based on the state, so all that block must be "synchronized", ie, protected
from running multiple time in separate threads.