acq align autofocus: run autofocus for _every_ grating x detector combination
We relied on the SR193 model that requires the focus position only for every grating once,
and every detector once. However, the hardware doesn't handle this properly, if the
focus is not set with the *first* detector for each grating.
Now that the SR193 driver overrides this behaviour, and just remembers the right focus
for each combination, it's more reliable to just autofocus on each grating x detector combination.

In practice, that usually means just one extra autofocus. In addition, as we optimise the grating
movements, it barely takes more time.