driver andorshrk: explicitely store the focus for each grating/output
The SR193 attempts to remember to the focus for the grating/output.
It is based on the model that the focus is based on the grating focus + output offset.
It might be correct, however the way the hardware stores the information is indirect,
by computing the focus values based on the focus positions. It's very opaque, and more
importantly, if doing the calibration on just the output 2, it's _impossible_ to
convince the hardware about keeping the right information.

Maybe the hardware could one day be made more clever (by relying only on the actual grating/output used),
but anyway, for all the current hardware already in production, it's much easier to
solve on the side of Odemis.

Also "improve" the simulator to mimick the current flawed behaviour of the hardware.

Also add a test case to check that focus is reset to the expected position when
focusing on the second output primarily.