gui plugin: sanitize Layout()
There was lots of confusion on the Layout()/Fit() and Refresh()/Update().
It caused some buttons to not be properly placed.

=> Use Layout() after new elements are shown/hidden (or added/removed),
so that they are all placed properly.

Fit() is only needed to adapt the size of the Window. So only at the beginning.

The Refresh()/Update() in general are not needed, because Fit() and Layout()
already trigger them.

Also fix the progress bar  being always displayed, even when the plugin has never used it.