Ensure that all strings with unicode characters are unicode strings
Eventually, in Python 3, all strings will be unicode.
However, for now, by default, strings are not unicode, and if they contain a unicode
character, they are utf-8 encoded (because the file itself is utf-8 encoded).

If we directly output such standard string to a file, it turns out to work fine (as long
as the file is also supposed to be utf-8 encoded). However, doing anything more fancy
on the string goes wrong. In particular, showing it in wxPython fails.

=> Convert all the strings containing non-ascii characters to unicode.
(I might have missed some though)

This fixes a bug on the SECOM when the fine alignment is run, and succeeds.
The GUI used to fail displaying the rotation. (and kept saying it's running the fine alignment)