streakcam: export tiff minor update
When reading older files saved with Odemis, the image might have higher dimensions
with CTZ=1. When reading the MD Odemis will add MD_TIME_LIST and MD_WL_LIST,
which one value. When exporting the data again, this was raising an issue.
MD_TIME_LIST and MD_WL_LIST are only saved now, if there are actually
higher dimensions of the image.