odemisd: Fix persistent metadata overridden by default metadata
If both persistent metadata and default metadata are available,
the persistent one should be picked. The code tried to do that,
but failed, and was doing the opposite.

Also add a test case for this.

Also fix a bug that would sometimes prevent setting some properties if both
persistent and default property existed for one of them only.

Also fix a bug preventing to start the backend without root permissions.
It's handy for just validating a file, or checking the backend status.

Also don't update the settings when only validating a file.