acq stream: simplify AutoBC -> recompute image logic
When AutoBC is enabled, we first display (quickly) the image with the
previous histogram, and once the new histogram is computed, we update
the intensity range, and the image.

We used to request for an image update when the histogram was updated,
or anything that changes the way the AutoBC computes the intensityRange.
Instead, now we just states that a change of intensityRange should
trigger an image update.

Until now, when A a new image arrived, the code path would look like:
 - _onNewData()
  -> _shouldUpdateHistogram()
  -> _shouldUpdateImage()
 - _updateImage()
  -> _projectXY2RGB()
   -> _getDisplayIRange()
    -> set intensityRange
 - _updateHistogram()
  -> sleep(0.25s)
  -> _shouldUpdateImage()
 - _updateImage()
  -> _projectXY2RGB()
   -> _getDisplayIRange()
    -> set intensityRange

Now we do:
 - _onNewData()
  -> _shouldUpdateHistogram()
  -> _shouldUpdateImage()
 - _updateImage()
  -> _projectXY2RGB()
   -> _getDisplayIRange()
 - _updateHistogram()
  -> _recomputeIntensityRange()
   -> _shouldUpdateImage()
 - _updateImage()
  -> _projectXY2RGB()
   -> _getDisplayIRange()

So the behaviour is slightly different. In some cases, it will reduce the computation,
and in other cases, it will slightly trigger more often an update of the image with the
updated intensityRange.

One main advantage is that this fixes the projections, which were not updated when
the histogram was changed due to moving along the Z index.