align autofocus: speed a little the autofocus by using the focus found in previous grating/detector
As the focus only varies a little bit between grating/detector,
it's not needed to check the entire focus range for each grating/detector.
So now we pass to the autofocus the focus from the previous grating/detector
as a "good_focus", and it will be compared to the current value, normally
stored from a previous autofocus session.

On the tests, it seems to work fine, although barely saving anytime (because changing the grating
takes the most time by far), it avoids the focus axis from moving too much.