gui util widgets AxisConnector: If a move caused no change in position, the current position should be set again
In some (rare) cases, a move request may be accepted by the hardware, but ends up causing
no change in position. As the widget was only updated when the position is updated,
the widget would not reflect anymore the current position (but just showed the requested position).

=> At the end of the move, also update the widget with the current position.
In the majority of the cases, this is going to duplicate the widget update,
but that ensures that in the corner cases, the position is still correctly reported.

In particular, that fixes the case of changing the spectrograph wavelength with a
mirror grating. The component silently ignores the move, so we need to report it.