util perpendicular_distance: Fix division by zero when segment is just a single point
Need to pay special attention when the segment is actually just a single point,
as it causes the norm to be null.

In practice, this caused an error on rare occasion when the user was selecting a
spectrum line (via the overlay) on the SPARC viewer.

11:05:13 ERROR   main           : Traceback (most recent call last):
  File "odemis\gui\comp\overlay\world.py", line 1149, in on_motion
  File "odemis\gui\comp\overlay\world.py", line 935, in on_motion
  File "odemis\gui\comp\overlay\base.py", line 947, in _on_motion
  File "odemis\gui\comp\overlay\base.py", line 885, in get_hover
  File "odemis\util\__init__.py", line 160, in perpendicular_distance
ZeroDivisionError: float division by zero