gui cont streams: When showing metadata, also insert ellipsis in lists
It's hard to balance between showing _all_ the metadata, and not showing gigantically too big data.
So limit the data to 2500, which slightly bigger than most CCD dimensions.
Therefore, when there is metadata value for each pixel, it's still entirely shown.

Now we enforce the limit to the lists as well (eg, MD_WL_LIST or MD_TIME_LIST).