util spot: avoid warning when using numpy.linalg.lstsq()
With numpy 1.11, it spits the following warning on the spirt use (and the GUI interprets it as an error).

FutureWarning: `rcond` parameter will change to the default of machine precision times ``max(M, N)`` wher
e M and N are the input matrix dimensions.
To use the future default and silence this warning we advise to pass `rcond=None`, to keep using the old, explicitly pass `rcond=-1`.
  ic, jc = numpy.linalg.lstsq(numpy.vstack((w * a, w * b)).T, w * c)[0]