acq align autofocus: revert autofocus spectrometer speed-up by using previous value
Revert commit 6e6eb8a486 (speed a little the autofocus by using the focus found in previous grating/detector).
It works in many cases, but causes some corner cases to break.
One particular example is the test case with two detectors.
One detector is not actually affected by the focus (because it's fake).
It causes the found focus to be incorrect, and the next one to also be wrong.

The speed improvement was so tiny that it's not worth to causes extra
issues.