acq align autofocus: If all the images are similar, don't move the focus
In case the focus has not effect on the image (eg, it's a fake detector, or there is
no signal), the autofocus would pick the lowest focus position.
So, such cases would cause the focus actuator to go way out-of-focus,
and reach the lowest limit.
=> Detect such cases, and just stay at the current position.

Also improve the log messages by reporting when there was not enough
data to find confidently the best focus position.