bug reporter: also store "lsusb -v" output
This can be handy to debug USB issues.

Also generalise the code to call external functions.
This has the extra benefit of handling timeout for all functions, and running them faster.
Also this avoids a failure whent the back-end was not running.

Also simplify the utf-8 handling, as json automatically converts everything to utf-8 anyway.