gui xrc: fix SPARCv2 align mode buttons (still) sometimes failing to be placed
Commit b9c4dcb283 (fix SPARC2 align FIBER button not showing) tried to make the
button placement work, but the code still had some issues in some cases.

The problem is that the grid sizer doesn't want to place two objects at the
same place. It's clever enough to not complain if trying to place the same object twice
at the same position. However, as the buttons were assigned positions by default,
a position could already be used when "moving" the buttons (and only later would
that original button be moved too).
This caused the streak cam GUI to not start.

=> Assigned default positions which are never used (eg, 5, N).
All the buttons will either be destroyed, or moved before ever being displayed.