streakcam and sim streak cam: fix calculation of wl range
The pixelsize was calculated by multiplying with the magnification. However, correct would be to divide by the magnification.
The pxsize will be larger for demagnification (aka covering a larger wl range per px). As we save the demagnification as values between 0 and 1, and magnification with values > 1, we need to divide by the magnification instead of multiplying.
* E.g. pxsize = 1um / 0.467 = 2.14 (demagnifying) > pxsize = 1um * 1 (no change in mag)