gui comp viewport: TwoDViewPorts don't need to listen to selected_pixel or selected_line anymore
Now that there is projection, whenever selected_pixel or selected_line are changed,
the projection is updated, and the viewports detect it.
So no need to explicitly listen to these VAs anymore.

Also move the connection of the MarkingLineOverlay to the viewport,
so the tab doesn't need to care, and it's the right stream which is always connected
(in the very unlikely case that there are multiple spectrum streams).