gui comp viewport: merge TimeSpectrumViewport and ChronographViewport
They have the same goal. ChronographViewport was initialy written for showing
the latest counts (of a single value) over time.
By adapting it to use the standard get_time_range, and making sure the streams
pass the proper metadata, it's able to show any type of chronograph.

In order to replicate exactly the old behaviour, also use MD_DT_NORMAL to indicate
that the data is normalized, and update the axis label to "count per second".
We leave the special +-5% margin code for all the data, as it's not a bad idea,
even for standard chronograms.

Also make sure to use that viewport (aka Monochromator) when using a time-correlator
on the SPARC. This fixes the time data not being displayed.