gui cont streams: only add streams to the *visible* views
When adding a stream with add_to_view=True, instead of adding it all
the views, only add it to the visible views.
It only changes something on the analysis tab, which doesn't always show
all the views (ie, visible_views < views). In such case, when adding a stream,
we should never need to add it to hidden views, as there is no way they will
be displayed (until loading another file).
That avoids doing too much (especially now that we have a lot of views), and
especially, this avoids creating DataProjections for SpectrumStreams which are
not supporting time (and so the chronograph view is hidden).