gui comp miccanvas: make TwoDPlotCanvas respect "flip" argument
It used to take the flip argument and... set it to 0, and pass it.
Obviously, the better way is to make sure the callers which don't want flipped
image just don't ask for it.

Most probably that code was originally for the temporal-spectrum display, and
now the projection already compute the image in the "correct" direction, so
no flip is required anymore.