gui comp viewport: TemporalSpectrumViewport takes care itself of connecting the mark lines
The MarkingLineOverlay used to take care of connecting the selected position to
two VAs, one for each axis. That code is used only once, as LineSpectrumViewport
need a different behaviour anyway. Moreover, every time the stream changed, the
connection had to be updated. However, that meant at the very first connection,
the canvas might not yet have any data, and drawing would fail.

=> Change TemporalSpectrumViewport to work the same way as LineSpectrumViewport:
it listens to the selected position by the overlay, and does the right thing,
depending on the current stream shown.

In practice, this means that it's now only one way: if selected_wavelength or selected_time
is changed, the marking line will not be updated. However, that's not useful,
especially because right now there is only one way for the user to change these values,
which is via the overlay.

Also, revert the MarkingLineOverlay draw code to only draw when the canvas has ranges.
That's actually a proxy to check whether the canvas has data. Ideally, we could just use
canvas.data, but curently that doesn't work with TwoDPlotCanvas. For PlotCanvas, that
means the caller should make sure to set the ranges (although that canvas works without
them).