gui util img: round back very tiny values to 0
Due to floating point errors, when computing the tick for a legend,
if the 0 is at the middle, it's not always exactly 0, but value 1e-16 times
smaller than the range.
For other numbers, this error is hidden when displaying the number,
but with 0, it's displayed as just a very very tiny number which needs
"e-N" notation.
That's quite ugly => detect it and change it back to "pure" 0, so that
it shows up nicely.