dataio hdf5: do not save a state for every value in DimensionScale*
_h5svi_set_state() tries to be clever, and automatically extends the state passed
if the underlying metadata has a dimension. This is correct in case we want to store
separate metadata for each C. However, when saving the scales, there is always just one,
for the whole data, so there should be only one state.

This was especially bad with MD_TIME_LIST, as it can be quite long (ie, 65536 values),
and the attributes in HDF5 are limited in size. This caused the entire file to not
being saved.

=> use a trick to convince _h5svi_set_state() that it should save just a single value.