acq path: add a error log if we detect that an axis is not referenced
In theory, it should never happen, as all the axes are referenced at initialisation.
In practice, sometimes an axis is not referenced. The most common reason is that there
is an issue with the hardware (eg, the reference switch doesn't work, or the axis doesn't
move). It could also be due to a later referencing which was canceled, in which case,
the position *might* still be correct.

We already report warning in the backend, but the user never see these ones.
So, also detect such condition in the optical path manager. At this point, we know
that it's very probably an error, so adequatly use the ERROR level.