acq stream sync SEMTemporalMDStream: fix shape and metadata of the final DataArrays
The folding of the array shape was incorrect: it attempted to put the T dimension at the C place,
and also the order was incorrect, so YX was mixed up in T.

Also the metadata was not properly computed:
TIME_LIST was not shorten as the data is, so afterwards Odemis would not trust it.
POS and PIXEL_SIZE were incorrectly computed => just use the standard function.
It's better to use the original DWELL_TIME from the detector, instead of overriding it with what
we hope it was.