util dataio: don't try to display a crazy amount of streams if the data has extra dimension
When the data has extra dimension, we break it into subdata.
However, the GUI doesn't handle any large number of streams. 10 is already big, and 30 is probably the maximum bearable.
So if for some reason this extra dimension is long (eg, a spectrum acquisition with the wrong shape or metadata),
let's not automatically try to show everything and freeze the GUI. => only show the first 10.