plugin spike_remove: clean-up and tiny new features
Disable the Save button when there is nothing to save.
Don't try to run on Temporal(Spectrum) data, altough it's a SpectrumStream too.

Clarify some comments.
Remove some attributes which weren't used or redundant.
Fix the author name.
Also ensure that the default window size is big enough, and that the content
is fitting at init.