support SPARC detectors with roles numbered
In order to support systems with many CCDs/spectrometers, instead of using explicit roles
such as "spectrometer" and "spectrometer-integrated", we now support roles such as "spectrometer0".
This is mostly done by using regex matching for "spectrometer.*", "ccd.*", and "sp-ccd*"

=> update the MD updater
=> update the OPM (which already supported regex matching)
=> extend gui.conf.data to support regex matching
This last one is the most extensive change, as now, all the lookups must be using
a special function get_hw_config().