gui conf file: encode unicode strings as UTF-8 and read them back too
All the filenames and paths are unicode (in the GUI).
However, we used to not care about unicode strings, and store them as-is in the .ini file.
Actually, Python would tend to do the right thing (ie, encode them to UTF-8) when writting
the file *but* it would read them back as standard strings. When such UTF-8 string would
contain a non-ascii character, converting it to unicode would fail.
Escentially, it could give this type of error at GUI start-up:

os.path.join(u"a", u"µ".encode("utf-8"))
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/lib/python2.7/posixpath.py", line 73, in join
    path += '/' + b
UnicodeDecodeError: 'ascii' codec can't decode byte 0xc2 in position 1: ordinal not in range(128)

=> Explicitly convert back every strings into unicode.