driver pigcs: new suspend_mode for closed-loop actuators
The newest E861 allow to turn off the sensors, but we actually don't typically want to
use that new feature.
=> disable it by default, and allow to use it.

Note that the current users of the C867 will have the default behaviour changed.
To restore the old behaviour they will need to add typically this line to the microscope file.
suspend_mode: {"sx": "full", "sy": "full"}