acq stream: simplify code assuming that .raw is a list
We used to have .raw either a list or a tuple. If it was a tuple, it would mean that
the real raw was a DataArrayShadow, and that .raw was the "intermediate" raw data for a
set of tiles. This is gone. Now .raw is a list, either of DataArrays or of DataArrayShadows.

Quite a few places can be simplified.