driver actuator LinearActuator: fix referencing between moves
Need to do the same thing for relative and absolute moves => share a function

When moving to _ref_start, don't do a recursive call to doMoveAbs(), as it could
easily lead to an infinite recursion. Just directly call the child.

Fix computation of the default ref_start, which needs to be offset.

Also simplify the moveRel(), to just do a moveRel(), not a moveAbs().