acq path: fix regex roles not matching any detector in getStreamDetector()
It still used a normal "==", while now all the role checks must use regex matching.

Also fix test cases to not worry about the spec-det-selector, if requesting a
"spectral" mode. As both ports of the spectrograph have "spectrometer.*",
it's fair that the OPM doesn't know which one to choose. Anyway, for such acquisition,
the mode should always be picked based on the stream, which works.