driver tmcm: fix high/low value confusion on do_axes
Sometimes the first value was considered high, while in other places, it
was considered low.
Also, revert the changes on led_do_prot, so that the values given in the
argument are actually the "reported value" of do_axes. It should
hopefully simplify a little bit the reading of the YAML file. Anyway, we
need do_axes, as we need to know how long to wait.