gui comp PlotViewport: always show the stream playing
Until now, whenever there was a stream added or removed, we would pick
the stream playing, as it's the most interesting one.
However, it's quite common that the user plays another stream afterwards.
In that case, the plot would still show the original stream.
Typical on SPARCs with multiple spectrometers.

=> listen to every stream.should_update, and whenever it's playing,
force that stream to be shown.