acq path: delay discarding non-affecting actuators from init to setPath()
With the regex support, there are several components that can match a single mode.
Some of them might be affected by an actuator, while others are not.
So it's not possible to just discard the actuator all the time.
This must be done once we know which detector is actually involved, which can only be
known in setPath().

Fixes the slit-in-big not being properly set for some spectral streams.