Bug fix: ar bg was not applied to image when loading a new file and reading the previous bg
While running the updateImage method for AR it calls AngleResolved2Polar, which needs some time.
However, while still in this code, the background VA is updated and resets the polar_dict.
After AngleResolved2Polar, the updateImage writes again in this dict, but with the previous not
bg corrected image. Then updateImage is called again for the changed background VA. However,
the code will now find the previous not bg corrected image in the dict and does not update
the image for the bg as requested.
Fix: Make a copy of the link. When the polar_dict is overwritten/cleared, only the copy is
modified in _project2polar with the new image. Thus when the updateImage is called and thus
_project2Polar again, the polar_dict is still empty. If polar_dict is not modified in between the
link is still the same and polar_dict has the same image as the previous copied link is pointing to.