acq path: setPath() accepts optional argument detector
For some alignment modes, just the mode is not sufficient to properly configure
the optical path when there are multiple detectors that could be used.
In such case, the caller can pass the detector (because it knows which one it actually wants).

Also add extra tests cases for the 4 spectrometers SPARC, to test not only this new feature,
but also the support of regex roles.