gui comp miccanvas NavigableBarPlotCanvas: don't fail to compute hrange if data has length 1
In case there is a single value, the for loop would be entirely skipped, and the
iright variable was never assigned... but used afterwards.

=> always initialise the values

Also fix the handling of the threshold: it would only crop the data if at least one
crop would be small. Instead, we want to crop on each side independently, and only
if it's big enough.